PlayRegisterの軽量化
