+++ src/types/PitchType.ts
@@ -1,0 +1,2 @@
+export type PitchType = 'rise' | 'drop' | 'cut' | 'changeup' | 'chenrai' | 'slider' | 'unknown';
+
